+++ tests/python/openassetio/hostAPI/test_manager.py
@@ -471,7 +471,8 @@ class Test_Manager_register:
 
 class Test_Manager_createContext:
 
-    def test_context_is_created_with_expected_properties(self, manager, mock_manager_interface):
+    def test_context_is_created_with_expected_properties(
+            self, manager, mock_manager_interface, mock_host_session):
 
         state_a = managerAPI.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_a
@@ -482,7 +483,7 @@ class Test_Manager_createContext:
         assert context_a.retention == Context.kTransient
         assert context_a.managerState is state_a
         assert context_a.locale is None
-        mock_manager_interface.mock.createState.assert_called_once()
+        mock_manager_interface.mock.createState.assert_called_once_with(mock_host_session)
 
     def test_when_called_with_parent_then_props_copied_and_createState_called_with_parent_state(
             self, manager, mock_manager_interface, mock_host_session):
@@ -496,7 +497,7 @@ class Test_Manager_createContext:
         mock_manager_interface.mock.reset_mock()
 
         state_b = managerAPI.ManagerStateBase()
-        mock_manager_interface.mock.createState.return_value = state_b
+        mock_manager_interface.mock.createChildState.return_value = state_b
 
         context_b = manager.createContext(parent=context_a)
 
@@ -505,8 +506,9 @@ class Test_Manager_createContext:
         assert context_b.access == context_a.access
         assert context_b.retention == context_a.retention
         assert context_b.locale == context_b.locale
-        mock_manager_interface.mock.createState.assert_called_once_with(
-            mock_host_session, parentState=state_a)
+        mock_manager_interface.mock.createChildState.assert_called_once_with(
+            mock_host_session, state_a)
+        mock_manager_interface.mock.createState.assert_not_called()
 
 
 class Test_Manager_freezeContext: